fix(lseek): return int64_t offset, drop truncation to int (T3.27)
fileseek's int return was casting a 64-bit offset to a 32-bit value
on the way out. The kernel-side path is now int64_t end-to-end —
the syscall dispatch already passed -lseek through an int64_t code
path, so this just removes the silent narrowing.

+++ kernel/file.c
@@ -104,7 +104,7 @@ int filestat(struct file *f, struct stat *st) {
     return f->ip->ops->stat(f->ip, st);
 }
 
-int fileseek(struct file *f, int64_t off, int whence) {
+int64_t fileseek(struct file *f, int64_t off, int whence) {
     if (f->type == FD_PIPE) return -ESPIPE;
     if (f->type != FD_INODE) return -ESPIPE;
     // Character devices are not seekable.
@@ -120,7 +120,7 @@ int fileseek(struct file *f, int64_t off, int whence) {
     int64_t newoff = base + off;
     if (newoff < 0) return -EINVAL;
     f->off = (uint64_t)newoff;
-    return (int)newoff;
+    return newoff;
 }
 
 int fileioctl(struct file *f, int cmd, unsigned long arg) {
